fix: feedback styles and mobile view
