--- desk/src/pages/TicketAgent/TicketAgent.vue
@@ -1,136 +1,0 @@
-<template>
-  <div v-if="ticket.doc?.name" class="flex-1">
-    <TicketHeader :viewers="viewers" />
-    <div class="h-full flex overflow-hidden">
-      <div class="flex-1 flex flex-col">
-        <!-- Tabs & Communication Area -->
-        <TicketActivityPanel />
-      </div>
-
-      <!-- Sidepanel with Resizer -->
-      <TicketSidebar />
-    </div>
-  </div>
-</template>
-
-<script setup lang="ts">
-import TicketActivityPanel from "@/components/ticket-agent/TicketActivityPanel.vue";
-import TicketHeader from "@/components/ticket-agent/TicketHeader.vue";
-import TicketSidebar from "@/components/ticket-agent/TicketSidebar.vue";
-import { useActiveViewers } from "@/composables/realtime";
-import { useTicket } from "@/composables/useTicket";
-import { ticketsToNavigate } from "@/composables/useTicketNavigation";
-import { globalStore } from "@/stores/globalStore";
-import {
-  ActivitiesSymbol,
-  AssigneeSymbol,
-  Customizations,
-  CustomizationSymbol,
-  RecentSimilarTicketsSymbol,
-  Resource,
-  TicketContactSymbol,
-  TicketSymbol,
-} from "@/types";
-import { createResource, toast } from "frappe-ui";
-import { computed, onBeforeUnmount, onMounted, provide, watch } from "vue";
-import { useRoute } from "vue-router";
-const { $socket } = globalStore();
-
-const props = defineProps({
-  ticketId: {
-    type: String,
-    required: true,
-  },
-});
-const route = useRoute();
-
-const ticketComposable = computed(() => useTicket(props.ticketId));
-const ticket = computed(() => ticketComposable.value.ticket);
-const customizations: Resource<Customizations> = createResource({
-  url: "helpdesk.helpdesk.doctype.hd_ticket.api.get_ticket_customizations",
-  cache: ["HD Ticket", "customizations"],
-  auto: true,
-});
-
-provide(TicketSymbol, ticket);
-
-provide(
-  AssigneeSymbol,
-  computed(() => ticketComposable.value.assignees)
-);
-provide(
-  TicketContactSymbol,
-  computed(() => ticketComposable.value.contact)
-);
-provide(
-  CustomizationSymbol,
-  computed(() => customizations)
-);
-provide(
-  RecentSimilarTicketsSymbol,
-  computed(() => ticketComposable.value.recentSimilarTickets)
-);
-provide(
-  ActivitiesSymbol,
-  computed(() => ticketComposable.value.activities)
-);
-
-const viewerComposable = computed(() => useActiveViewers(ticket.value.name));
-const viewers = computed(
-  () => viewerComposable.value.currentViewers[props.ticketId] || []
-);
-const { startViewing, stopViewing } = viewerComposable.value;
-
-// handling for faster navigation between tickets
-watch(
-  () => route.params.ticketId,
-  (newTicketId, oldTicketId) => {
-    if (newTicketId === oldTicketId) return;
-
-    if (oldTicketId) stopViewing(oldTicketId as string);
-    startViewing(newTicketId as string);
-  },
-  { immediate: true }
-);
-
-type TicketUpdateData = {
-  ticket_id: string;
-  user: string;
-  field: string;
-  value: string;
-};
-
-onMounted(() => {
-  // startViewing(props.ticketId);
-
-  ticketsToNavigate.update({
-    params: {
-      ticket: props.ticketId,
-      current_view: route.query.view as string,
-    },
-  });
-  ticketsToNavigate.reload();
-  ticket.value.markSeen.reload();
-
-  $socket.on("ticket_update", (data: TicketUpdateData) => {
-    if (data.ticket_id === ticket.value?.name) {
-      // Notify the user about the update
-      toast.info(`User ${data.user} updated ${data.field} to ${data.value}`);
-    }
-  });
-});
-
-onBeforeUnmount(() => {
-  stopViewing(props.ticketId);
-});
-</script>
-
-<style>
-.breadcrumbs button {
-  background-color: inherit !important;
-  &:hover,
-  &:focus {
-    background-color: inherit !important;
-  }
-}
-</style>
